fix: Cooked analysis tool

+++ src/common/task.py
@@ -3,8 +3,8 @@ from dataclasses import dataclass
 @dataclass
 class Task:
     task_name:str
-    wcet:int
-    period:int
+    wcet:float
+    period:float
     component_id:str
     priority:int|None
